examples/a8-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
redis==5.0.0
psycopg2-binary==2.9.9
httpx==0.27.0
